Update qt conversion to use runtime polymorphic conversion
