+++ platform/qt/qt.cmake
@@ -64,6 +64,10 @@ include_directories(
 add_library(qmapboxgl SHARED
     platform/qt/include/qmapbox.hpp
     platform/qt/include/qmapboxgl.hpp
+    platform/qt/src/qt_conversion.cpp
+    platform/qt/src/qt_conversion.hpp
+    platform/qt/src/qt_geojson.cpp
+    platform/qt/src/qt_geojson.hpp
     platform/qt/src/qmapbox.cpp
     platform/qt/src/qmapboxgl.cpp
     platform/qt/src/qmapboxgl_p.hpp
